pyproject.toml
[project]
name = "handwriting-generator"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch==1.13.1",
    "pytorch_lightning-1.8.6",
    "numpy==1.24.0",
    "matplotlib==3.6.2",
    "click==8.1.3",
    "tqdm==4.64.1",
    "jupyter==1.0.0",
]

[build-system]
requires = ["setuptools >= 40.6.0", "wheel"]
build-backend = "setuptools.build_meta"

# Black-compatible settings for isort
# See https://black.readthedocs.io/en/stable/compatible_configs.html
[tool.isort]
profile = "black"
